+++ packages/ricos-content/web/src/consts.ts
@@ -132,6 +132,7 @@ export const RICOS_NODE_TYPE_TO_DATA_FIELD = {
   [Node_Type.VIDEO]: 'videoData',
   [Node_Type.POLL]: 'pollData',
   [Node_Type.TABLE]: 'tableData',
+  [Node_Type.TABLE_CELL]: 'tableCellData',
   [Node_Type.PARAGRAPH]: 'paragraphData',
   [Node_Type.LIST_ITEM]: 'paragraphData',
   [Node_Type.HEADING]: 'headingData',