+++ packages/ricos-content/web/src/converters/html/fromHtml/rich-text/parser.ts
@@ -11,6 +11,9 @@ import {
   backgroundStyleToTextHighlight,
   aToLink,
   textToText,
+  spanToMention,
+  spanToSpoiler,
+  aToAnchor,
 } from '../core/rules';
 import { Rule } from '../core/models';
 
@@ -27,8 +30,11 @@ export default flow(
     lToList,
     hToHeading,
     aToLink,
+    aToAnchor,
     strongEmUToDecoration,
     colorStyleToTextColor,
+    spanToMention,
+    spanToSpoiler,
     backgroundStyleToTextHighlight,
   ])
 );